Move `Cases` tests into `cases.rs`
